fix: remove circular dependency between utils and deepmerge files
